requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-PyMongo==2.3.0
Flask-WTF==1.1.1
pymongo==4.5.0
bcrypt==4.0.1
Pillow==10.0.0
python-dotenv==1.0.0
email-validator==2.0.0.post2
werkzeug==2.3.7
